fix(diag): time dispatches with QPC; stop the PERF probe measuring itself
winmm proxy LIVE-VERIFIED on The Adventures of Elliot (UE 5.4): 180/180 exports
forwarded to the real System32 winmm, lazily at T+1.2s on a game thread after
DllMain returned rather than under the loader lock, then a clean UE5_Init
(GObjects/GNames/GWorld found, 326364 objects, name sanity 10/10). All four
proxy flavours now work.

The automatic PERF records then earned their keep on their first outing by
exposing two defects in themselves -- which a hand-run measurement session would
very likely have shrugged past:

  PERF Copy CE Field: wall 57.7 ms . dispatcher busy 93 ms (161.2%)
     . top: walk_instance 0ms/128x max 15ms

161% busy, with the breakdown contradicting the total. Two independent causes:

1. GetTickCount64 was the wrong clock. Its ~15.6 ms granularity floors every
   sub-tick dispatch to zero, so 128 walk_instance calls summed to "0 ms" while
   one that happened to straddle a tick read 15 ms -- an artefact of tick
   alignment, not a measurement, and sub-millisecond commands are precisely the
   population this exists to measure. Sense now times with
   QueryPerformanceCounter and accumulates microseconds, reporting fractional ms
   on the wire (total_ms / max_ms / last_ms became doubles, and the C# model
   with them).

2. The probe was measuring itself. `busy` came from the global total_busy_ms
   delta, which includes the probe's own opening get_diagnostics (~93 ms), while
   the per-command ranking already excluded it -- hence 93 ms of busy against a
   57.7 ms operation whose top row showed 0 ms. Busy is now summed from the
   per-command deltas, so the percentage and the breakdown agree by construction
   rather than by coincidence.

Both pinned by regression tests carrying the real numbers from the log. A
genuine >100% remains possible and meaningful -- the two-connection lane split
can have two dispatchers busy at once -- so the figure is deliberately not
capped.

Verification: all 4 proxies + main DLL build clean; 2896 tests green (+3).
Still to do: re-read the PERF lines with the fixed clock. The pre-fix numbers
understate sub-ms commands and overstate short operations, so the multipipe
Phase 1 decision should wait for fresh ones.

+++ ui/UE5DumpUI/Models/DiagnosticsResult.cs
@@ -11,11 +11,15 @@ public sealed class DiagnosticsCommandEntry
 {
     public string Cmd     { get; init; } = "";
     public long   Count   { get; init; }
-    public long   TotalMs { get; init; }
+    /// <summary>Fractional milliseconds. The DLL times dispatches with QPC and
+    /// reports microseconds scaled to ms, because most pipe commands are
+    /// SUB-millisecond — an integer here would round the interesting ones to
+    /// zero, which is exactly the bug the first live run exposed.</summary>
+    public double TotalMs { get; init; }
     /// <summary>Worst single dispatch — the head-of-line spike a user feels as a
     /// frozen UI.</summary>
-    public long   MaxMs   { get; init; }
-    public long   LastMs  { get; init; }
+    public double MaxMs   { get; init; }
+    public double LastMs  { get; init; }
     public double AvgMs   { get; init; }
 
     /// <summary>Share of all dispatcher-busy time owned by this command. Set by the
@@ -81,7 +85,8 @@ public sealed class DiagnosticsResult
 {
     public long   UptimeMs        { get; init; }
     public long   TotalDispatches { get; init; }
-    public long   TotalBusyMs     { get; init; }
+    /// <summary>Fractional milliseconds — see <see cref="DiagnosticsCommandEntry.TotalMs"/>.</summary>
+    public double TotalBusyMs     { get; init; }
     /// <summary>Fraction of wall-clock a dispatcher was occupied. High here plus a
     /// lagging UI is the case FOR non-blocking dispatch; low here says the lag is
     /// somewhere else and Phase 1 would not help.</summary>
